pyproject.toml
[tool.poetry]
name = "etl_pipeline"
version = "0.1.0"
description = ""
authors = ["erik-tastepe <erik.tastepe@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
pyspark = "^3.1.1"
pandas = "^1.2.3"
PyYAML = "^5.4.1"
psycopg2 = "^2.8.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^3.9.0"
pydocstyle = "^5.1.1"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
